+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update `disabled` state on `Contribution` component.
 - Feature: Update style on `ContributionCard` component.
 
 ## [21.4.0] - 2018-08-06